requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
graphene==3.2.2
graphql-core==3.2.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
flask-cors==4.0.0
email-validator==2.1.0
gunicorn==21.2.0
requests==2.31.0 